Grib

+++ H2OOP/Models/PremiumUser.cs
@@ -3,6 +3,6 @@ public class PremiumUser : User
 {
     public PremiumUser(string? name) : base(name)
     {
-        MaximumAllowedBooks = int.MaxValue;
+        MaximumAllowedBooks = ushort.MaxValue;
     }
 }
